extracting_yaml_data/requirements.txt
pyyaml>=6.0.1
requests>=2.31.0
rich>=13.6.0
